+++ src/components/index.ts
@@ -5,5 +5,6 @@ import Text from './text'
 import Title from './title'
 import TextField from './form/text-field'
 import Image from './image'
+import IconButton from './icon-button'
 
-export { Button, Icon, PasswordField, Text, TextField, Title, Image }
+export { Button, Icon, IconButton, PasswordField, Text, TextField, Title, Image }