+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - [update]: Readme file
 - [added]: Support for loader on buttons
 - [added]: Cogs and home icon
+- [added]: IconButton component